+++ operator/src/main/java/oracle/kubernetes/operator/webhooks/utils/GsonBuilderUtils.java
@@ -4,12 +4,14 @@
 package oracle.kubernetes.operator.webhooks.utils;
 
 import java.time.OffsetDateTime;
+import java.util.Map;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import oracle.kubernetes.operator.helpers.GsonOffsetDateTime;
 import oracle.kubernetes.operator.webhooks.model.AdmissionReview;
 import oracle.kubernetes.operator.webhooks.model.ConversionReviewModel;
+import oracle.kubernetes.weblogic.domain.model.DomainResource;
 
 public class GsonBuilderUtils {
 
@@ -31,6 +33,26 @@ public class GsonBuilderUtils {
 
   public static String writeAdmissionReview(AdmissionReview admissionReview) {
     return getGsonBuilder().toJson(admissionReview, AdmissionReview.class);
+  }
+
+  public static DomainResource readDomain(String resourceName) {
+    return getGsonBuilder().fromJson(resourceName, DomainResource.class);
+  }
+
+  public static String writeDomain(DomainResource domain) {
+    return getGsonBuilder().toJson(domain,DomainResource.class);
+  }
+
+  public static Map<String, Object> writeDomainToMap(DomainResource domain) {
+    return readMap(getGsonBuilder().toJson(domain, DomainResource.class));
+  }
+
+  public static String writeMap(Map map) {
+    return getGsonBuilder().toJson(map, Map.class);
+  }
+
+  public static Map<String, Object> readMap(String map) {
+    return getGsonBuilder().fromJson(map, Map.class);
   }
 
   private static Gson getGsonBuilder() {
